src/agentic_platform/agent/pydanticai_agent/requirements.txt
pydantic-ai-slim[bedrock]
fastapi
uvicorn
boto3
pyjwt
cryptography
requests